clients/IssueListItem: add highlights to PR titles
