+++ clients/apps/web/src/components/Dashboard/IssueListItem.tsx
@@ -13,10 +13,11 @@ import { IssueReadWithRelations } from 'polarkit/api/types'
 import {
   IssueActivityBox,
   IssueListItemDecoration,
+  generateMarkdownTitle,
 } from 'polarkit/components/Issue'
 import { PolarTimeAgo, PrimaryButton } from 'polarkit/components/ui'
 import { getCentsInDollarString, githubIssueUrl } from 'polarkit/utils'
-import React, { ChangeEvent, useState } from 'react'
+import { ChangeEvent, useState } from 'react'
 import PledgeNow from '../Pledge/PledgeNow'
 import IconCounter from './IconCounter'
 import IssueLabel, { LabelSchema } from './IssueLabel'
@@ -210,54 +211,6 @@ const IssueListItem = (props: {
       )}
     </>
   )
-}
-
-export const generateMarkdownTitle = (
-  title: string,
-): React.ReactElement | React.ReactElement[] => {
-  const matches: RegExpMatchArray[] = []
-  for (const m of title.matchAll(/`([^`]*)`/g)) {
-    matches.push(m)
-  }
-
-  if (matches.length === 0) {
-    return <>{title}</>
-  }
-
-  let i = 0
-  let offset = 0
-  const nodes: React.ReactElement[] = []
-  const matchCount = matches.length
-
-  for (const match of matches) {
-    if (match.index === undefined) {
-      continue
-    }
-
-    i += 1
-    if (offset < match.index) {
-      nodes.push(
-        <React.Fragment key={`0-${i}`}>
-          {title.substring(offset, match.index)}
-        </React.Fragment>,
-      )
-    }
-
-    nodes.push(
-      <span key={`1-${i}`} className="rounded-md bg-gray-100 py-0.5 px-1.5">
-        {match[1]}
-      </span>,
-    )
-    offset = match.index + match[0].length
-    if (i === matchCount) {
-      nodes.push(
-        <React.Fragment key={`3-${i}`}>
-          {title.substring(offset, title.length)}
-        </React.Fragment>,
-      )
-    }
-  }
-  return nodes
 }
 
 export default IssueListItem
